Qt4/Mac requires a type to avoid an "ambiguous argument type" error.
git-svn-id: http://svn.osgeo.org/qgis/trunk/qgis@4256 c8812cc2-4d05-0410-92ff-de0c093fc19c

+++ src/qgsmapcanvas.cpp
@@ -2876,7 +2876,7 @@ bool QgsMapCanvas::writeXML(QDomNode & layerNode, QDomDocument & doc)
   // Iterate over layers in zOrder
   // Call writeXML() on each
   QDomElement projectLayersNode = doc.createElement("projectlayers");
-  projectLayersNode.setAttribute("layercount", mCanvasProperties->layers.size());
+  projectLayersNode.setAttribute("layercount", qulonglong(mCanvasProperties->layers.size()));
 
   std::list < QString >::iterator li = mCanvasProperties->zOrder.begin();
   while (li != mCanvasProperties->zOrder.end())
